Add files via upload
Question 1:
Create a program which have three variables, all type int, should be called day, month, and year.
Write a program that prompts the user to enter values in day, month, and year. The program then
stores the data in a variables, and finally prints it out.

Question 2:
Implement a class named Point to represent a point with x and y coordinates. The class contains:
 Two attributes x and y that represent coordinates.
 Methods to get and set these attributes.
 A constructor that creates a point with specified coordinates.
A method named distance that returns the distance from the current point to another point whose x
and y coordinates are sent as arguments.

Question 3:
Make a menu of the following operations using two dimensional array of size m x n. You
should use user-defined functions which accept 2-D array A, and its size m and n as
arguments. The options are:
 To input elements into matrix of size m x n
 To display elements of matrix of size m x n
 Sum of all elements of matrix of size m x n
 To display row-wise sum of matrix of size m x n
 To display column-wise sum of matrix of size m x n
 To create transpose of matrix of size n x m